pyproject.toml
[project]
name = "tidal-discovery-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click",
    "tidalapi",
    "pylast",
    "python-dotenv",
    "google-genai>=1.61.0",
]
